Core/Maps: Oops guid is a ulong not a uint.

+++ Source/Game/Maps/Map.cs
@@ -3253,7 +3253,7 @@ namespace Game.Maps
             {
                 do
                 {
-                    var loguid = result.Read<uint>(0);
+                    var loguid = result.Read<ulong>(0);
                     var respawnTime = result.Read<long>(1);
 
                     CreatureData cdata = Global.ObjectMgr.GetCreatureData(loguid);
